refactor code to satisfy codacy
* also improved code formatting for a file or two
