+++ src/test/java/de/uhd/ifi/se/decision/management/jira/extraction/gitclient/TestGetFeatureBranchCommits.java
@@ -3,11 +3,8 @@ package de.uhd.ifi.se.decision.management.jira.extraction.gitclient;
 import de.uhd.ifi.se.decision.management.jira.extraction.impl.GitClientImpl;
 import org.eclipse.jgit.lib.Ref;
 import org.eclipse.jgit.revwalk.RevCommit;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
-import java.io.File;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -15,24 +12,16 @@ import static org.junit.Assert.assertEquals;
 
 public class TestGetFeatureBranchCommits extends TestSetUpGit {
 
-	final String repoBaseDirectory;
-	final String uri;
-	GitClientImpl testGitClient;
+	private final String repoBaseDirectory;
+	private final String uri;
+	private GitClientImpl testGitClient;
 
-	String featureBranch = "featureBranch";
-	String expectedFirstCommitMessage = "First message";
+	private String featureBranch = "featureBranch";
+	private String expectedFirstCommitMessage = "First message";
 
 	public TestGetFeatureBranchCommits() {
 		repoBaseDirectory = super.getRepoBaseDirectory();
 		uri = super.getRepoUri();
-	}
-
-	@Before
-	public void setUp() {
-	}
-
-	@After
-	public void tearDown() {
 	}
 
 	@Test
