+++ src/test/java/de/uhd/ifi/se/decision/management/jira/extraction/versioncontrol/TestGitDecXtract.java
@@ -8,8 +8,8 @@ import org.junit.Test;
 import java.util.List;
 
 public class TestGitDecXtract extends TestSetUpGit {
-	final String uri;
-	GitDecXtract gitDecX;
+	private final String uri;
+	private GitDecXtract gitDecX;
 
 	public TestGitDecXtract() {
 		uri = super.getRepoUri();
